* master:
  Remove /* inline */
  Remove __fastcall
  Fix x64 compilation
  Change code style
  Fix warnings
  int i --> intptr_t I
  Bookmarks.cpp: Revert temp changes
  *.vcxproj: Compile in build\ directory

Conflicts:
	src/NetBox/WinSCPDialogs.cpp
	src/core/Configuration.cpp
	src/core/Exceptions.cpp
	src/core/Exceptions.h
	src/core/FileMasks.cpp
	src/core/FileMasks.h
	src/core/FtpFileSystem.cpp
	src/core/Queue.cpp
	src/core/Queue.h
	src/core/SessionData.cpp
	src/core/SessionData.h
	src/core/SessionInfo.cpp
	src/core/SessionInfo.h
	src/core/Terminal.cpp
	src/core/Terminal.h
	src/windows/GUITools.cpp
	src/windows/GUITools.h

+++ src/NetBox/FarPlugin.cpp
@@ -2330,7 +2330,7 @@ void TFarKeyBarTitles::ClearFileKeyBarTitles()
 }
 //------------------------------------------------------------------------------
 void TFarKeyBarTitles::ClearKeyBarTitle(TFarShiftStatus ShiftStatus,
-  int FunctionKeyStart, int FunctionKeyEnd)
+  intptr_t FunctionKeyStart, intptr_t FunctionKeyEnd)
 {
   if (!FunctionKeyEnd)
   {
@@ -2343,7 +2343,7 @@ void TFarKeyBarTitles::ClearKeyBarTitle(TFarShiftStatus ShiftStatus,
 }
 //------------------------------------------------------------------------------
 void TFarKeyBarTitles::SetKeyBarTitle(TFarShiftStatus ShiftStatus,
-  int FunctionKey, const UnicodeString & Title)
+  intptr_t FunctionKey, const UnicodeString & Title)
 {
   assert(FunctionKey >= 1 && FunctionKey <= static_cast<intptr_t>(LENOF(FKeyBarTitles.Titles)));
   wchar_t ** Titles = NULL;
