+++ src/core/RemoteFiles.h
@@ -65,7 +65,7 @@ public:
   void Log(TTerminalIntf * Terminal, const wchar_t * Title);
 
   intptr_t GetCount() const;
-  const TRemoteToken * Token(int Index) const;
+  const TRemoteToken * Token(intptr_t Index) const;
 
 private:
   typedef std::vector<TRemoteToken> TTokens;
@@ -88,7 +88,7 @@ private:
   TDateTime FModification;
   TDateTime FLastAccess;
   TRemoteToken FGroup;
-  Integer FIconIndex;
+  intptr_t FIconIndex;
   Boolean FIsSymLink;
   TRemoteFile * FLinkedFile;
   TRemoteFile * FLinkedByFile;
@@ -103,7 +103,7 @@ private:
   UnicodeString FTypeName;
 
 public:
-  int GetAttr();
+  intptr_t GetAttr();
   bool GetBrokenLink();
   bool GetIsDirectory() const;
   TRemoteFile * GetLinkedFile();
@@ -119,7 +119,7 @@ public:
   void SetRights(TRights * Value);
   UnicodeString GetFullFileName() const;
   bool GetHaveFullFileName() const;
-  int GetIconIndex() const;
+  intptr_t GetIconIndex() const;
   UnicodeString GetTypeName();
   bool GetIsHidden();
   void SetIsHidden(bool Value);
