+++ src/core/SessionData.cpp
@@ -349,7 +349,7 @@ void TSessionData::Assign(TSessionDataIntf * Source)
       (static_cast<TSessionData *>(Source))->SetBug(static_cast<TSshBug>(Index),
           GetBug(static_cast<TSshBug>(Index)));
     }
-    for (unsigned int Index = 0; Index < LENOF(FSFTPBugs); ++Index)
+    for (intptr_t Index = 0; Index < static_cast<intptr_t>(LENOF(FSFTPBugs)); ++Index)
     {
       // PROPERTY(SFTPBug[(TSftpBug)Index]);
       (static_cast<TSessionData *>(Source))->SetSFTPBug(static_cast<TSftpBug>(Index),
@@ -1686,13 +1686,13 @@ bool TSessionData::GetUsesSsh()
   return (GetFSProtocol() < fsFTP);
 }
 //------------------------------------------------------------------------------
-void TSessionData::SetCipher(int Index, TCipher Value)
+void TSessionData::SetCipher(intptr_t Index, TCipher Value)
 {
   assert(Index >= 0 && Index < CIPHER_COUNT);
   SET_SESSION_PROPERTY(Ciphers[Index]);
 }
 //------------------------------------------------------------------------------
-TCipher TSessionData::GetCipher(int Index) const
+TCipher TSessionData::GetCipher(intptr_t Index) const
 {
   assert(Index >= 0 && Index < CIPHER_COUNT);
   return FCiphers[Index];
@@ -1701,10 +1701,13 @@ TCipher TSessionData::GetCipher(int Index) const
 void TSessionData::SetCipherList(const UnicodeString & Value)
 {
   bool Used[CIPHER_COUNT];
-  for (int C = 0; C < CIPHER_COUNT; C++) { Used[C] = false; }
+  for (intptr_t C = 0; C < CIPHER_COUNT; C++)
+  {
+    Used[C] = false;
+  }
 
   UnicodeString CipherStr;
-  int Index = 0;
+  intptr_t Index = 0;
   UnicodeString Value2 = Value;
   while (!Value2.IsEmpty() && (Index < CIPHER_COUNT))
   {
@@ -1737,13 +1740,13 @@ UnicodeString & TSessionData::GetCipherList() const
   return FCipherListStr;
 }
 //------------------------------------------------------------------------------
-void TSessionData::SetKex(int Index, TKex Value)
+void TSessionData::SetKex(intptr_t Index, TKex Value)
 {
   assert(Index >= 0 && Index < KEX_COUNT);
   SET_SESSION_PROPERTY(Kex[Index]);
 }
 //------------------------------------------------------------------------------
-TKex TSessionData::GetKex(int Index) const
+TKex TSessionData::GetKex(intptr_t Index) const
 {
   assert(Index >= 0 && Index < KEX_COUNT);
   return FKex[Index];
@@ -1755,7 +1758,7 @@ void TSessionData::SetKexList(const UnicodeString & Value)
   for (int K = 0; K < KEX_COUNT; K++) { Used[K] = false; }
 
   UnicodeString KexStr;
-  int Index = 0;
+  intptr_t Index = 0;
   UnicodeString Value2 = Value;
   while (!Value2.IsEmpty() && (Index < KEX_COUNT))
   {
@@ -3100,7 +3103,7 @@ TSessionData * TStoredSessionList::FindSame(TSessionData * Data)
   return Result;
 }
 //------------------------------------------------------------------------------------
-int TStoredSessionList::IndexOf(TSessionData * Data)
+intptr_t TStoredSessionList::IndexOf(TSessionData * Data)
 {
   for (intptr_t Index = 0; Index < GetCount(); ++Index)
   {
